+++ EasyPost/Models/API/Payload.cs
@@ -1,0 +1,13 @@
+namespace EasyPost.Models.API
+{
+    public class Payload : Event // A payload has the same schema as an Event
+    {
+        #region JSON Properties
+
+        #endregion
+
+        internal Payload()
+        {
+        }
+    }
+}